requirements/base.txt
django==5.1
python-environ==0.4.54
psycopg2-binary==2.9.9
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-spectacular==0.27.2